extensions/instances/embeddings/requirements.txt
pillow == 10.4.0
torch == 2.5.1
torchvision == 0.20.1
accelerate == 1.2.1
transformers == 4.44.0
openai-clip == 1.0.1
picteus-internal-extension-sdk
